fixing type miscomparison of uint64_t with int

+++ C_RPi_Andy.cpp
@@ -38,7 +38,7 @@ int main(int argc, char* argv[])
       if (sole->thresholdCross >= 3) {
         bool modeSwitch = false;
         for (size_t i = 2; i < thresholdTimes.size(); i++) {
-            if (thresholdTimes[i] - thresholdTimes[i - 2] <= 3) { // within 3 seconds
+            if (thresholdTimes[i] - thresholdTimes[i - 2] <= static_cast<uint64_t>(3)) { // within 3 seconds
                 modeSwitch = true;
                 break;
             }
